All fixed beside massive method args
Исправил все кроме:
1) У меня методы сортировки вызывались с разным количеством параметров. Но, не которые из них рекурсивны, поэтому нельзя было этого избежать. Я могу для стилистики добавить бесполезный аргумент с длиной массива там, где это не нужно, но это странно.
2) С башем пока не разбирался
