+++ lesson1/homework/src/com/hsai/home1/src/alghorithm/insert_Sort.java
@@ -1,0 +1,14 @@
+package alghorithm;
+public class insert_Sort{
+    public static void insertionSort(int[] array) {  
+        for (int i = 1; i < array.length; i++) {
+            int current = array[i];
+            int j = i - 1;
+            while(j >= 0 && current < array[j]) {
+                array[j+1] = array[j];
+                j--;
+            }
+            array[j+1] = current;
+        }
+    }
+}
