--- lesson1/homework/src/com/hsai/home1/src/insertSort/insert_Sort.java
@@ -1,14 +1,0 @@
-package insertSort;
-public class insert_Sort{
-    public static void insertionSort(int[] array) {  
-        for (int i = 1; i < array.length; i++) {
-            int current = array[i];
-            int j = i - 1;
-            while(j >= 0 && current < array[j]) {
-                array[j+1] = array[j];
-                j--;
-            }
-            array[j+1] = current;
-        }
-    }
-}
